Move deployment tests to another file
